Added base class `BaseTestCase` for shared `setUp` and `tearDown` code.
